Refactor config-related code
