+++ CMakeLists.txt
@@ -77,8 +77,6 @@ set(CORE_SOURCES
   src/character_status.cpp
   src/class.cpp
   src/command.cpp
-  src/config.cpp
-  src/config_def.cpp
   src/crafting.cpp
   src/ctrl_file.cpp
   src/ctrl_inventory.cpp
@@ -134,7 +132,6 @@ set(CORE_SOURCES
   src/random.cpp
   src/random_event.cpp
   src/set_item_info.cpp
-  src/set_option.cpp
   src/shop.cpp
   src/spec.cpp
   src/status_ailment.cpp
@@ -184,6 +181,11 @@ set(CORE_SOURCES
   src/lua_env/lua_class/lua_class.cpp
   src/lua_env/lua_class/lua_class_item.cpp
   src/lua_env/lua_class/lua_class_position.cpp
+
+  src/config/config.cpp
+  src/config/config_def.cpp
+  src/config/set_option.cpp
+  src/config/create_config_screen.cpp
 
   src/dialog/dialog.cpp
   src/dialog/dialog_data.cpp
