+++ src/blending.cpp
@@ -3,7 +3,7 @@
 #include "activity.hpp"
 #include "audio.hpp"
 #include "character.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "db_item.hpp"
 #include "draw.hpp"
 #include "elona.hpp"
