+++ src/building.cpp
@@ -5,7 +5,7 @@
 #include "audio.hpp"
 #include "calc.hpp"
 #include "character.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "ctrl_file.hpp"
 #include "db_item.hpp"
 #include "draw.hpp"
