+++ src/card.cpp
@@ -1,7 +1,7 @@
 #include "card.hpp"
 #include "audio.hpp"
 #include "character.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "elona.hpp"
 #include "i18n.hpp"
 #include "input.hpp"
