+++ src/config/config.cpp
@@ -4,14 +4,14 @@
 #include <functional>
 #include <stdexcept>
 #include <string>
-#include "elona.hpp"
-#include "hcl.hpp"
-#include "range.hpp"
-#include "snail/android.hpp"
-#include "snail/application.hpp"
-#include "snail/touch_input.hpp"
-#include "snail/window.hpp"
-#include "variables.hpp"
+#include "../elona.hpp"
+#include "../hcl.hpp"
+#include "../range.hpp"
+#include "../snail/android.hpp"
+#include "../snail/application.hpp"
+#include "../snail/touch_input.hpp"
+#include "../snail/window.hpp"
+#include "../variables.hpp"
 
 
 namespace
