+++ src/config/config.hpp
@@ -2,13 +2,13 @@
 
 #include <iostream>
 #include <string>
+#include "../elona.hpp"
+#include "../filesystem.hpp"
+#include "../lib/noncopyable.hpp"
+#include "../log.hpp"
+#include "../snail/window.hpp"
+#include "../thirdparty/ordered_map/ordered_map.h"
 #include "config_def.hpp"
-#include "elona.hpp"
-#include "filesystem.hpp"
-#include "lib/noncopyable.hpp"
-#include "log.hpp"
-#include "snail/window.hpp"
-#include "thirdparty/ordered_map/ordered_map.h"
