+++ src/crafting.cpp
@@ -3,7 +3,7 @@
 #include "audio.hpp"
 #include "calc.hpp"
 #include "character.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "draw.hpp"
 #include "i18n.hpp"
 #include "input.hpp"
