+++ src/draw.cpp
@@ -1,7 +1,7 @@
 #include "draw.hpp"
 #include <cmath>
 #include "character.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "db_chara_chip.hpp"
 #include "db_item.hpp"
 #include "db_item_chip.hpp"
