+++ src/magic.cpp
@@ -7,7 +7,7 @@
 #include "calc.hpp"
 #include "character.hpp"
 #include "character_status.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "ctrl_file.hpp"
 #include "db_item.hpp"
 #include "debug.hpp"
