+++ src/random_event.cpp
@@ -6,7 +6,7 @@
 #include "calc.hpp"
 #include "character.hpp"
 #include "character_status.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "dmgheal.hpp"
 #include "draw.hpp"
 #include "event.hpp"
