+++ src/spec_load.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <map>
 #include <vector>
-#include "config_def.hpp"
+#include "config/config_def.hpp"
 #include "filesystem.hpp"
 #include "hcl.hpp"
 #include "optional.hpp"
