+++ src/talk_house_visitor.cpp
@@ -6,7 +6,7 @@
 #include "calc.hpp"
 #include "character.hpp"
 #include "character_status.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "i18n.hpp"
 #include "item.hpp"
 #include "itemgen.hpp"
