+++ src/tests/config_def.cpp
@@ -1,6 +1,6 @@
 #include "../thirdparty/catch2/catch.hpp"
 
-#include "../config_def.hpp"
+#include "../config/config_def.hpp"
 #include "../testing.hpp"
 #include "tests.hpp"
