+++ src/turn_sequence.cpp
@@ -10,7 +10,7 @@
 #include "character.hpp"
 #include "character_status.hpp"
 #include "command.hpp"
-#include "config.hpp"
+#include "config/config.hpp"
 #include "ctrl_file.hpp"
 #include "db_item.hpp"
 #include "debug.hpp"
